Fix 32-bit type mismatch

+++ libringbuffer/shm_types.h
@@ -20,7 +20,7 @@ struct shm_object {
 	int wait_fd[2];	/* fd for wait/wakeup */
 	char *memory_map;
 	int is_shadow;
-	size_t memory_map_size;
+	uint64_t memory_map_size;
 	size_t allocated_len;
 };
